+++ src/main/java/net/mohron/skyclaims/command/CommandIsland.java
@@ -25,6 +25,9 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import net.mohron.skyclaims.PluginInfo;
+import net.mohron.skyclaims.command.team.CommandInvite;
+import net.mohron.skyclaims.command.team.CommandKick;
+import net.mohron.skyclaims.command.team.CommandLeave;
 import net.mohron.skyclaims.command.user.CommandCreate;
 import net.mohron.skyclaims.command.user.CommandExpand;
 import net.mohron.skyclaims.command.user.CommandInfo;
@@ -47,6 +50,7 @@ import org.spongepowered.api.command.CommandSource;
 import org.spongepowered.api.command.args.CommandContext;
 import org.spongepowered.api.command.args.GenericArguments;
 import org.spongepowered.api.command.spec.CommandSpec;
+import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.service.pagination.PaginationList;
 import org.spongepowered.api.text.Text;
 import org.spongepowered.api.text.action.TextActions;
@@ -94,6 +98,9 @@ public class CommandIsland extends CommandBase {
         CommandCreate.register();
         CommandExpand.register();
         CommandInfo.register();
+        CommandInvite.register();
+        CommandKick.register();
+        CommandLeave.register();
         CommandList.register();
         CommandLock.register();
         CommandRegen.register();
@@ -106,7 +113,6 @@ public class CommandIsland extends CommandBase {
 
     public CommandResult execute(CommandSource src, CommandContext args) throws CommandException {
         List<Text> helpText = Lists.newArrayList();
-        boolean hasPerms = false;
 
         helpText.add(Text.of(
             TextColors.WHITE, "SkyClaims utilizes GriefPrevention for world protection and management. Go to ",
@@ -121,7 +127,6 @@ public class CommandIsland extends CommandBase {
                 TextColors.DARK_GRAY, " - ",
                 TextColors.DARK_GREEN, CommandCreate.HELP_TEXT
             ));
-            hasPerms = true;
         }
 
         if (src.hasPermission(Permissions.COMMAND_HOME)) {
@@ -130,7 +135,6 @@ public class CommandIsland extends CommandBase {
                 TextColors.DARK_GRAY, " - ",
                 TextColors.DARK_GREEN, CommandHome.HELP_TEXT
             ));
-            hasPerms = true;
         }
 
         if (src.hasPermission(Permissions.COMMAND_EXPAND)) {
@@ -140,7 +144,6 @@ public class CommandIsland extends CommandBase {
                 TextColors.DARK_GRAY, " - ",
                 TextColors.DARK_GREEN, CommandExpand.HELP_TEXT
             ));
-            hasPerms = true;
         }
 
         if (src.hasPermission(Permissions.COMMAND_INFO)) {
@@ -150,7 +153,33 @@ public class CommandIsland extends CommandBase {
                 TextColors.DARK_GRAY, " - ",
                 TextColors.DARK_GREEN, CommandInfo.HELP_TEXT
             ));
-            hasPerms = true;
+        }
+
+        if (src.hasPermission(Permissions.COMMAND_INVITE)) {
+            helpText.add(Text.of(
+                TextColors.AQUA, Text.builder("is invite").onClick(TextActions.runCommand("/is invite")),
+                TextColors.GRAY, " [user]",
+                TextColors.GRAY, " [privilege]",
+                TextColors.DARK_GRAY, " - ",
+                TextColors.DARK_GREEN, CommandInvite.HELP_TEXT
+            ));
+        }
+
+        if (src.hasPermission(Permissions.COMMAND_KICK)) {
+            helpText.add(Text.of(
+                TextColors.AQUA, Text.builder("is kick").onClick(TextActions.suggestCommand("/is kick")),
+                TextColors.GRAY, " [user]",
+                TextColors.DARK_GRAY, " - ",
+                TextColors.DARK_GREEN, CommandKick.HELP_TEXT
+            ));
+        }
+
+        if (src.hasPermission(Permissions.COMMAND_LEAVE)) {
+            helpText.add(Text.of(
+                TextColors.AQUA, Text.builder("is leave").onClick(TextActions.runCommand("/is leave")),
+                TextColors.DARK_GRAY, " - ",
+                TextColors.DARK_GREEN, CommandLeave.HELP_TEXT
+            ));
         }
 
         if (src.hasPermission(Permissions.COMMAND_LIST)) {
@@ -161,7 +190,6 @@ public class CommandIsland extends CommandBase {
                 TextColors.DARK_GRAY, " - ",
                 TextColors.DARK_GREEN, CommandList.HELP_TEXT
             ));
-            hasPerms = true;
         }
 
         if (src.hasPermission(Permissions.COMMAND_LOCK)) {
@@ -171,7 +199,6 @@ public class CommandIsland extends CommandBase {
                 TextColors.DARK_GRAY, " - ",
                 TextColors.DARK_GREEN, CommandLock.HELP_TEXT
             ));
-            hasPerms = true;
         }
 
         if (src.hasPermission(Permissions.COMMAND_REGEN)) {
@@ -181,7 +208,6 @@ public class CommandIsland extends CommandBase {
                 TextColors.DARK_GRAY, " - ",
                 TextColors.DARK_GREEN, CommandRegen.HELP_TEXT
             ));
-            hasPerms = true;
         }
 
         if (src.hasPermission(Permissions.COMMAND_RESET)) {
@@ -191,7 +217,6 @@ public class CommandIsland extends CommandBase {
                 TextColors.DARK_GRAY, " - ",
                 TextColors.DARK_GREEN, CommandReset.HELP_TEXT
             ));
-            hasPerms = true;
         }
 
         if (src.hasPermission(Permissions.COMMAND_SET_BIOME)) {
@@ -202,7 +227,6 @@ public class CommandIsland extends CommandBase {
                 TextColors.DARK_GRAY, " - ",
                 TextColors.DARK_GREEN, CommandSetBiome.HELP_TEXT
             ));
-            hasPerms = true;
         }
 
         if (src.hasPermission(Permissions.COMMAND_SET_HOME)) {
@@ -211,7 +235,6 @@ public class CommandIsland extends CommandBase {
                 TextColors.DARK_GRAY, " - ",
                 TextColors.DARK_GREEN, CommandSetHome.HELP_TEXT
             ));
-            hasPerms = true;
         }
 
         if (src.hasPermission(Permissions.COMMAND_SET_SPAWN)) {
@@ -220,7 +243,6 @@ public class CommandIsland extends CommandBase {
                 TextColors.DARK_GRAY, " - ",
                 TextColors.DARK_GREEN, CommandSetSpawn.HELP_TEXT
             ));
-            hasPerms = true;
         }
 
         if (src.hasPermission(Permissions.COMMAND_SPAWN)) {
@@ -230,7 +252,6 @@ public class CommandIsland extends CommandBase {
                 TextColors.DARK_GRAY, " - ",
                 TextColors.DARK_GREEN, CommandSpawn.HELP_TEXT
             ));
-            hasPerms = true;
         }
 
         if (src.hasPermission(Permissions.COMMAND_LOCK)) {
@@ -240,17 +261,20 @@ public class CommandIsland extends CommandBase {
                 TextColors.DARK_GRAY, " - ",
                 TextColors.DARK_GREEN, CommandUnlock.HELP_TEXT
             ));
-            hasPerms = true;
         }
 
-        if (hasPerms) {
-            PaginationList.builder()
-                .title(Text.of(TextColors.AQUA, NAME, " Help"))
-                .padding(Text.of(TextColors.AQUA, TextStyles.STRIKETHROUGH, "-"))
-                .contents(helpText)
-                .sendTo(src);
+        if (helpText.size() > 1) {
+            if (src instanceof Player) {
+                PaginationList.builder()
+                    .title(Text.of(TextColors.AQUA, NAME, " Help"))
+                    .padding(Text.of(TextColors.AQUA, TextStyles.STRIKETHROUGH, "-"))
+                    .contents(helpText)
+                    .sendTo(src);
+            } else {
+                helpText.forEach(src::sendMessage);
+            }
         } else {
-            src.sendMessage(Text.of(NAME + " " + VERSION));
+            src.sendMessage(Text.of(NAME, " ", VERSION));
         }
 
         return CommandResult.success();