+++ src/main/java/net/mohron/skyclaims/command/team/CommandInvite.java
@@ -18,32 +18,51 @@
 
 package net.mohron.skyclaims.command.team;
 
+import com.google.common.collect.ImmutableList;
 import net.mohron.skyclaims.command.CommandBase;
+import net.mohron.skyclaims.command.CommandIsland;
+import net.mohron.skyclaims.permissions.Options;
 import net.mohron.skyclaims.permissions.Permissions;
+import net.mohron.skyclaims.team.Invite;
+import net.mohron.skyclaims.team.PrivilegeType;
+import net.mohron.skyclaims.world.Island;
 import org.spongepowered.api.command.CommandException;
 import org.spongepowered.api.command.CommandResult;
-import org.spongepowered.api.command.CommandSource;
 import org.spongepowered.api.command.args.CommandContext;
 import org.spongepowered.api.command.args.GenericArguments;
 import org.spongepowered.api.command.spec.CommandSpec;
 import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.entity.living.player.User;
+import org.spongepowered.api.service.pagination.PaginationList;
 import org.spongepowered.api.text.Text;
+import org.spongepowered.api.text.format.TextColors;
+import org.spongepowered.api.text.format.TextStyles;
 
-public class CommandInvite extends CommandBase {
+import javax.annotation.Nonnull;
+
+public class CommandInvite extends CommandBase.IslandCommand {
 
     public static final String HELP_TEXT = "used to invite players to your island or list your pending invites.";
+    private static final Text LIST = Text.of("list");
     private static final Text USER = Text.of("user");
+    private static final Text PRIVILEGE = Text.of("privilege");
 
     public static CommandSpec commandSpec = CommandSpec.builder()
         .permission(Permissions.COMMAND_INVITE)
-        .arguments(GenericArguments.optional(GenericArguments.user(USER)))
+        .arguments(
+            GenericArguments.optional(GenericArguments.seq(
+                GenericArguments.user(USER),
+                GenericArguments.optional(GenericArguments.enumValue(PRIVILEGE, PrivilegeType.class))
+            )),
+            GenericArguments.optional(GenericArguments.literal(LIST, "list"))
+        )
         .description(Text.of(HELP_TEXT))
         .executor(new CommandInvite())
         .build();
 
     public static void register() {
         try {
+            CommandIsland.addSubCommand(commandSpec, "invite");
             PLUGIN.getGame().getCommandManager().register(PLUGIN, commandSpec);
             PLUGIN.getLogger().debug("Registered command: CommandInvite");
         } catch (UnsupportedOperationException e) {
@@ -52,22 +71,69 @@ public class CommandInvite extends CommandBase {
         }
     }
 
-    @Override
-    public CommandResult execute(CommandSource src, CommandContext args) throws CommandException {
-        if (!(src instanceof Player)) {
-            throw new CommandException(Text.of("You must be a player to run this command!"));
-        }
-        Player player = (Player) src;
-        User user = (User) args.getOne(USER).orElse(null);
+    @Override public CommandResult execute(@Nonnull Player player, @Nonnull Island island, @Nonnull CommandContext args) throws CommandException {
+        User user = args.<User>getOne(USER).orElse(null);
+        PrivilegeType type = args.<PrivilegeType>getOne(PRIVILEGE).orElse(PrivilegeType.MEMBER);
 
-        if (user == null) {
-            //TODO List invites
+        if (type == PrivilegeType.NONE) {
+            throw new CommandException(
+                Text.of(TextStyles.ITALIC, "What kind of invite is ", TextStyles.RESET, type.toText(), TextStyles.ITALIC, "?"));
+        }
+
+        if (user == null || args.hasAny(LIST)) {
+            PaginationList.builder()
+                .title(Text.of(TextColors.AQUA, "Invite List"))
+                .header(getIslandLimit(player))
+                .padding(Text.of(TextColors.AQUA, TextStyles.STRIKETHROUGH, "-"))
+                .contents(PLUGIN.getInviteService().getInviteText(player).isEmpty()
+                    ? ImmutableList.of(Text.of(TextColors.RED, "You have no pending invites!"))
+                    : PLUGIN.getInviteService().getInviteText(player)
+                )
+                .sendTo(player);
+        } else if (player.equals(user)) {
+            throw new CommandException(Text.of(TextColors.RED, "You cannot invite yourself!"));
+        } else if (island.getPrivilegeType(user) == type) {
+            throw new CommandException(Text.of(
+                island.getPrivilegeType(user).format(user.getName()), TextColors.RED, " is already a ", type.toText(), TextColors.RED, "!"
+            ));
+        } else if (type == PrivilegeType.MEMBER && !island.isManager(player) || !island.getOwnerUniqueId().equals(player.getUniqueId())) {
+            throw new CommandException(Text.of(
+                TextColors.RED, "You do not have permission to send ", type == PrivilegeType.OWNER ? "an " : "a ",
+                type.toText(),
+                TextColors.RED, " invite for ",
+                island.getName(),
+                TextColors.RED, "!"
+            ));
+        } else if (island.getPrivilegeType(user) == PrivilegeType.MEMBER || island.getPrivilegeType(user) == PrivilegeType.MANAGER) {
+            player.sendMessage(Text.of(
+                type.format(user.getName()),
+                TextColors.GREEN, " has been changed from a ",
+                island.getPrivilegeType(user).toText(),
+                " to a ",
+                type.toText(), "."
+            ));
+            island.removeMember(user);
+            island.addMember(user, type);
         } else {
-            //TODO Send invite to online player or save for login
+            Invite invite = Invite.builder()
+                .island(island)
+                .sender(player)
+                .receiver(user)
+                .privilegeType(type)
+                .build();
+            if (PLUGIN.getInviteService().inviteExists(invite)) {
+                throw new CommandException(Text.of(TextColors.RED, "Invite already exists!"));
+            } else {
+                invite.send();
+                player.sendMessage(Text.of(TextColors.GREEN, "Island invite sent to ", type.format(user.getName()), TextColors.GREEN, "."));
+            }
         }
 
-        throw new CommandException(Text.of("Command not yet implemented"));
+        return CommandResult.success();
+    }
 
-        //return CommandResult.success();
+    private Text getIslandLimit(Player player) {
+        int limit = Options.getMaxIslands(player.getUniqueId());
+        return (limit < 1) ? null : Text.of(TextColors.GRAY, "You may join up to ", TextColors.LIGHT_PURPLE, limit, TextColors.GRAY, " islands.");
     }
 }