+++ src/main/java/net/mohron/skyclaims/command/user/CommandInfo.java
@@ -24,6 +24,7 @@ import net.mohron.skyclaims.command.CommandIsland;
 import net.mohron.skyclaims.command.argument.Argument;
 import net.mohron.skyclaims.permissions.Options;
 import net.mohron.skyclaims.permissions.Permissions;
+import net.mohron.skyclaims.team.PrivilegeType;
 import net.mohron.skyclaims.util.CommandUtil;
 import net.mohron.skyclaims.world.Island;
 import org.spongepowered.api.command.CommandException;
@@ -43,6 +44,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.UUID;
 import java.util.function.Consumer;
+import java.util.stream.Collectors;
 
 public class CommandInfo extends CommandBase {
 
@@ -65,6 +67,55 @@ public class CommandInfo extends CommandBase {
             e.printStackTrace();
             PLUGIN.getLogger().error("Failed to register command: CommandInfo");
         }
+    }
+
+    public CommandResult execute(CommandSource src, CommandContext args) throws CommandException {
+        List<Island> islands = Lists.newArrayList();
+
+        if (src instanceof Player && !args.hasAny(ISLAND)) {
+            islands.add(Island.get(((Player) src).getLocation())
+                .orElseThrow(() -> new CommandException(Text.of(TextColors.RED, "You must be on an island to use this command.")))
+            );
+        } else {
+            Collection<UUID> islandIds = args.getAll(ISLAND);
+            islandIds.forEach(i -> Island.get(i).ifPresent(islands::add));
+            if (islands.size() > 1) {
+                return listIslands();
+            }
+        }
+
+        List<Text> infoText = islands.stream()
+            .map(island -> Text.of(
+                (src instanceof Player) ? getAdminShortcuts(src, island) : Text.EMPTY,
+                TextColors.YELLOW, "Name", TextColors.WHITE, " : ", island.getName(),
+                getLocked(island), Text.NEW_LINE,
+                TextColors.YELLOW, "Members", TextColors.WHITE, " : ", getMembers(island), Text.NEW_LINE,
+                TextColors.YELLOW, "Size", TextColors.WHITE, " : ", TextColors.LIGHT_PURPLE, island.getWidth(),
+                TextColors.GRAY, "x", TextColors.LIGHT_PURPLE, island.getWidth(), Text.NEW_LINE,
+                TextColors.YELLOW, "Entities", TextColors.WHITE, " : ", getEntities(island), Text.NEW_LINE,
+                TextColors.YELLOW, "Spawn", TextColors.WHITE, " : ", getSpawn(island), Text.NEW_LINE,
+                TextColors.YELLOW, "Created", TextColors.WHITE, " : ", TextColors.GRAY, island.getDateCreated(), Text.NEW_LINE,
+                TextColors.YELLOW, "Last Active", TextColors.WHITE, " : ", TextColors.GRAY, island.getDateLastActive(), Text.NEW_LINE,
+                TextColors.YELLOW, "UUID", TextColors.WHITE, " : ", TextColors.GRAY, island.getUniqueId(), Text.NEW_LINE,
+                (island.getClaim().isPresent()) ? Text.of(
+                    TextColors.YELLOW, "Claim", TextColors.WHITE, " : ", TextColors.GRAY, Text.builder(
+                        island.getClaimUniqueId().toString())
+                        .onClick(TextActions.executeCallback(CommandUtil.createCommandConsumer(src, "claiminfo",
+                            island.getClaimUniqueId().toString(),
+                            createReturnConsumer(src, island.getUniqueId().toString())
+                        )))
+                        .onHover(TextActions.showText(Text.of("Click here to check claim info.")))
+                ) : Text.EMPTY
+            ))
+            .collect(Collectors.toList());
+
+        PaginationList.builder()
+            .title(Text.of(TextColors.AQUA, "Island Info"))
+            .padding(Text.of(TextColors.AQUA, TextStyles.STRIKETHROUGH, "-"))
+            .contents(infoText)
+            .sendTo(src);
+
+        return CommandResult.success();
     }
 
     private static CommandResult listIslands() throws CommandException {
@@ -137,20 +188,16 @@ public class CommandInfo extends CommandBase {
     }
 
     private static Text getMembers(Island island) {
-        if (island.getMembers().isEmpty()) {
-            return Text.of(TextColors.GRAY, "None");
-        } else {
-            int i = 1;
-            Text members = Text.EMPTY;
-            for (String member : island.getMembers()) {
-                members = Text.join(members, Text.of(
-                    TextColors.BLUE, member, TextColors.GRAY,
-                    (i == island.getMembers().size()) ? "" : ", ")
-                );
-                i++;
-            }
-            return members;
+        List<Text> members = Lists.newArrayList();
+        members.add(PrivilegeType.OWNER.format(island.getOwnerName()));
+        for (String manager : island.getManagers()) {
+            members.add(PrivilegeType.MANAGER.format(manager));
         }
+        for (String member : island.getMembers()) {
+            members.add(PrivilegeType.MEMBER.format(member));
+        }
+        return Text.joinWith(Text.of(TextColors.GRAY, ", "), members);
+
     }
 
     private static Consumer<CommandSource> createReturnConsumer(CommandSource src, String arguments) {
@@ -164,57 +211,6 @@ public class CommandInfo extends CommandBase {
                 .build();
             src.sendMessage(returnCommand);
         };
-    }
-
-    public CommandResult execute(CommandSource src, CommandContext args) throws CommandException {
-        List<Text> infoText = Lists.newArrayList();
-        List<Island> islands = Lists.newArrayList();
-
-        if (src instanceof Player && !args.hasAny(ISLAND)) {
-            islands.add(Island.get(((Player) src).getLocation())
-                .orElseThrow(() -> new CommandException(Text.of(TextColors.RED, "You must be on an island to use this command.")))
-            );
-        } else {
-            Collection<UUID> islandIds = args.getAll(ISLAND);
-            islandIds.forEach(i -> Island.get(i).ifPresent(islands::add));
-            if (islands.size() > 1) {
-                return listIslands();
-            }
-        }
-
-        islands.forEach(island -> {
-            infoText.add(Text.of(
-                (src instanceof Player) ? getAdminShortcuts(src, island) : Text.EMPTY,
-                TextColors.YELLOW, "Name", TextColors.WHITE, " : ", island.getName(),
-                getLocked(island), Text.NEW_LINE,
-                TextColors.YELLOW, "Owner", TextColors.WHITE, " : ", TextColors.GOLD, island.getOwnerName(), Text.NEW_LINE,
-                TextColors.YELLOW, "Members", TextColors.WHITE, " : ", getMembers(island), Text.NEW_LINE,
-                TextColors.YELLOW, "Size", TextColors.WHITE, " : ", TextColors.LIGHT_PURPLE, island.getWidth(),
-                TextColors.GRAY, "x", TextColors.LIGHT_PURPLE, island.getWidth(), Text.NEW_LINE,
-                TextColors.YELLOW, "Entities", TextColors.WHITE, " : ", getEntities(island), Text.NEW_LINE,
-                TextColors.YELLOW, "Spawn", TextColors.WHITE, " : ", getSpawn(island), Text.NEW_LINE,
-                TextColors.YELLOW, "Created", TextColors.WHITE, " : ", TextColors.GRAY, island.getDateCreated(), Text.NEW_LINE,
-                TextColors.YELLOW, "Last Active", TextColors.WHITE, " : ", TextColors.GRAY, island.getDateLastActive(), Text.NEW_LINE,
-                TextColors.YELLOW, "UUID", TextColors.WHITE, " : ", TextColors.GRAY, island.getUniqueId(), Text.NEW_LINE,
-                (island.getClaim().isPresent()) ? Text.of(
-                    TextColors.YELLOW, "Claim", TextColors.WHITE, " : ", TextColors.GRAY, Text.builder(
-                        island.getClaimUniqueId().toString())
-                        .onClick(TextActions.executeCallback(CommandUtil.createCommandConsumer(src, "claiminfo",
-                            island.getClaimUniqueId().toString(),
-                            createReturnConsumer(src, island.getUniqueId().toString())
-                        )))
-                        .onHover(TextActions.showText(Text.of("Click here to check claim info.")))
-                ) : Text.EMPTY
-            ));
-        });
-
-        PaginationList.builder()
-            .title(Text.of(TextColors.AQUA, "Island Info"))
-            .padding(Text.of(TextColors.AQUA, TextStyles.STRIKETHROUGH, "-"))
-            .contents(infoText)
-            .sendTo(src);
-
-        return CommandResult.success();
     }
 
     private Text getLocked(Island island) {
@@ -255,7 +251,7 @@ public class CommandInfo extends CommandBase {
     }
 
     private Text getEntities(Island island) {
-        boolean showLimits = PLUGIN.getConfig().getEntityConfig().isLimitSpawning();
+        boolean limitSpawning = PLUGIN.getConfig().getEntityConfig().isLimitSpawning();
         int max = Options.getMaxSpawns(island.getOwnerUniqueId());
         int maxHostile = Options.getMaxHostileSpawns(island.getOwnerUniqueId());
         int hostile = island.getHostileEntities().size();
@@ -265,21 +261,21 @@ public class CommandInfo extends CommandBase {
         return Text.of(
             Text.of(TextActions.showText(Text.of(
                 TextColors.GRAY, "Hostile", TextColors.WHITE, " : ", TextColors.LIGHT_PURPLE, hostile,
-                (showLimits && maxHostile > 0)
+                (limitSpawning && maxHostile > 0)
                     ? Text.of(TextColors.GRAY, " (", TextColors.RED, maxHostile, TextColors.GRAY, ")")
                     : Text.EMPTY,
                 Text.NEW_LINE,
                 TextColors.GRAY, "Passive", TextColors.WHITE, " : ", TextColors.LIGHT_PURPLE, passive,
-                (showLimits && maxPassive > 0)
+                (limitSpawning && maxPassive > 0)
                     ? Text.of(TextColors.GRAY, " (", TextColors.RED, maxPassive, TextColors.GRAY, ")")
                     : Text.EMPTY
                 )),
-                TextColors.GRAY, "Living", TextColors.WHITE, ": ", TextColors.LIGHT_PURPLE, hostile + passive, TextColors.WHITE,
-                (showLimits && max > 0)
+                TextColors.GRAY, "Living", TextColors.WHITE, " : ", TextColors.LIGHT_PURPLE, hostile + passive, TextColors.WHITE,
+                (limitSpawning && max > 0)
                     ? Text.of(TextColors.GRAY, " (", TextColors.RED, max, TextColors.GRAY, ")")
                     : Text.EMPTY, ", "
             ),
-            TextColors.GRAY, "Item", TextColors.WHITE, ": ", TextColors.LIGHT_PURPLE, island.getItemEntities().size(), TextColors.WHITE, ", ",
-            TextColors.GRAY, "Tile", TextColors.WHITE, ": ", TextColors.LIGHT_PURPLE, island.getTileEntities().size());
+            TextColors.GRAY, "Item", TextColors.WHITE, " : ", TextColors.LIGHT_PURPLE, island.getItemEntities().size(), TextColors.WHITE, ", ",
+            TextColors.GRAY, "Tile", TextColors.WHITE, " : ", TextColors.LIGHT_PURPLE, island.getTileEntities().size());
     }
 }