+++ src/main/java/net/mohron/skyclaims/world/Island.java
@@ -20,7 +20,7 @@ package net.mohron.skyclaims.world;
 
 import com.flowpowered.math.vector.Vector3d;
 import com.flowpowered.math.vector.Vector3i;
-import com.google.common.collect.Sets;
+import com.google.common.collect.Lists;
 import me.ryanhamshire.griefprevention.api.claim.Claim;
 import me.ryanhamshire.griefprevention.api.claim.ClaimManager;
 import me.ryanhamshire.griefprevention.api.claim.ClaimResult;
@@ -30,6 +30,7 @@ import net.mohron.skyclaims.SkyClaims;
 import net.mohron.skyclaims.exception.CreateIslandException;
 import net.mohron.skyclaims.exception.InvalidRegionException;
 import net.mohron.skyclaims.permissions.Options;
+import net.mohron.skyclaims.team.PrivilegeType;
 import net.mohron.skyclaims.util.ClaimUtil;
 import net.mohron.skyclaims.world.region.IRegionPattern;
 import net.mohron.skyclaims.world.region.Region;
@@ -53,9 +54,9 @@ import org.spongepowered.api.world.World;
 import java.util.Collection;
 import java.util.Date;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -295,13 +296,46 @@ public class Island {
         return getWidth() == width;
     }
 
-    public Set<String> getMembers() {
-        Set<String> members = Sets.newHashSet();
+    public void addMember(User user) {
+        addMember(user, PrivilegeType.MEMBER);
+    }
+
+    public void addMember(User user, PrivilegeType type) {
+        switch (type) {
+            case OWNER:
+                UUID previousOwner = owner;
+                transfer(user);
+                getClaim().ifPresent(c -> c.addUserTrust(previousOwner, TrustType.MANAGER, PLUGIN.getCause()));
+                break;
+            case MANAGER:
+            case MEMBER:
+                getClaim().ifPresent(c -> c.addUserTrust(user.getUniqueId(), type.getTrustType(), PLUGIN.getCause()));
+                break;
+            case NONE:
+                getClaim().ifPresent(c -> c.removeUserTrust(user.getUniqueId(), type.getTrustType(), PLUGIN.getCause()));
+                break;
+        }
+    }
+
+    public void removeMember(User user) {
+        getClaim().ifPresent(c -> c.removeUserTrust(user.getUniqueId(), TrustType.NONE, PLUGIN.getCause()));
+    }
+
+    public List<String> getMembers() {
+        List<String> members = Lists.newArrayList();
         if (!getClaim().isPresent()) {
             return members;
         }
         for (UUID builder : getClaim().get().getUserTrusts(TrustType.BUILDER)) {
             members.add(getName(builder));
+        }
+        return members;
+    }
+
+    public List<String> getManagers() {
+        List<String> members = Lists.newArrayList();
+        if (!getClaim().isPresent()) {
+            return members;
         }
         for (UUID manager : getClaim().get().getUserTrusts(TrustType.MANAGER)) {
             members.add(getName(manager));
@@ -331,6 +365,22 @@ public class Island {
     public boolean isManager(User user) {
         return user.getUniqueId().equals(owner)
             || (getClaim().isPresent() && getClaim().get().isUserTrusted(user, TrustType.MANAGER));
+    }
+
+    public boolean isOwner(User user) {
+        return user.getUniqueId().equals(owner);
+    }
+
+    public PrivilegeType getPrivilegeType(User user) {
+        if (isOwner(user)) {
+            return PrivilegeType.OWNER;
+        } else if (isManager(user)) {
+            return PrivilegeType.MANAGER;
+        } else if (isMember(user)) {
+            return PrivilegeType.MEMBER;
+        } else {
+            return PrivilegeType.NONE;
+        }
     }
 
     public Collection<Player> getPlayers() {